Tests/LibWeb/Text/expected/HTML/document-open-after-completed-script-created-parser.txt

Iframe load 1 fired with ready state complete
Iframe load 2 fired with ready state complete
